Avoid exception under Cygwin
Cygwin is a hybrid between Windows and Unix.  It does not currently
provide ffmpeg as a package, so the user will have to rely on Windows
builds, such as the [Zeranoe](https://ffmpeg.zeranoe.com/builds/) ones.
Setting winNN can work with such an installation and a suitable path.

Issue: #26